made changes to naming of create redux action, changing linter errors/naming convention of actions, and removed the no data action creator
