+++ src/actions/create-redux-action/create-redux-action.js
@@ -1,0 +1,2 @@
+export const createReduxAction = actionType => data => ({ type: actionType, data });
+export default createReduxAction;
