+++ src/actions/create-redux-action/index.js
@@ -1,0 +1,1 @@
+export { default } from './create-redux-action';
