+++ src/actions/login-modal/login-modal-actions.spec.js
@@ -1,7 +1,6 @@
 import Actions from '../../reducers/actions';
 import { closeLoginModal, swapModal } from '.';
 
-
 describe('close login modal action', () => {
   const testActionType = Actions.login.closeLoginModal;
   it('should return a lambda that returns an object with the closeLoginModal action', () => {
