+++ src/actions/navbar/navbar-actions.js
@@ -1,5 +1,5 @@
 import Actions from '../../reducers/actions';
-import { reduxActionCreatorNoData } from '../redux-action-creator';
+import createReduxAction from '../create-redux-action';
 
-const setLoggedIn = reduxActionCreatorNoData(Actions.login.isLoggedIn);
+const setLoggedIn = createReduxAction(Actions.login.isLoggedIn);
 export default setLoggedIn;
