--- src/actions/redux-action-creator/index.js
@@ -1,1 +1,0 @@
-export * from './redux-action-creator';
