--- src/actions/redux-action-creator/redux-action-creator.js
@@ -1,2 +1,0 @@
-export const reduxActionCreator = actionType => data => ({ type: actionType, data });
-export const reduxActionCreatorNoData = actionType => () => ({ type: actionType });
