+++ src/actions/signup-modal/signup-modal-actions.spec.js
@@ -1,7 +1,6 @@
 import Actions from '../../reducers/actions';
 import { closeSignupModal, swapModal } from '.';
 
-
 describe('close login modal action', () => {
   const testActionType = Actions.login.closeSignupModal;
   it('should return a lambda that returns an object with the closeSignupModal action', () => {
