+++ src/containers/connected-login-modal/connected-login-modal.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import * as LoginModalActions from '../../actions/login-modal';
+import * as loginModalActions from '../../actions/login-modal';
 import loginModalComponent from '../../components/login-modal';
 
 const mapStateToProps = state => ({
@@ -9,7 +9,7 @@ const mapStateToProps = state => ({
 const mapDispatchToProps = dispatch =>
   bindActionCreators(
     {
-      ...LoginModalActions,
+      ...loginModalActions,
     },
     dispatch,
   );
