+++ src/containers/connected-pipeline/connected-request-review/connected-request-review.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import * as PipelineActions from '../../../actions/pipeline';
+import * as pipelineActions from '../../../actions/pipeline';
 import requestReviewComponent from '../../../pages/pipeline/request-review';
 
 const mapStateToProps = state => ({
@@ -12,7 +12,7 @@ const mapStateToProps = state => ({
 const mapDispatchToProps = dispatch =>
   bindActionCreators(
     {
-      ...PipelineActions,
+      ...pipelineActions,
     },
     dispatch,
   );
