+++ src/containers/connected-pipeline/connected-request-selection/connected-request-selection.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import * as PipelineActions from '../../../actions/pipeline';
+import * as pipelineActions from '../../../actions/pipeline';
 import requestSelectionComponent from '../../../pages/pipeline/request-selection';
 
 const mapStateToProps = state => ({
@@ -10,7 +10,7 @@ const mapStateToProps = state => ({
 const mapDispatchToProps = dispatch =>
   bindActionCreators(
     {
-      ...PipelineActions,
+      ...pipelineActions,
     },
     dispatch,
   );
