+++ src/containers/connected-signup-modal/connected-signup-modal.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import signupModalComponent from '../../components/signup-modal';
-import * as SignupModalActions from '../../actions/signup-modal';
+import * as signupModalActions from '../../actions/signup-modal';
 
 const mapStateToProps = state => ({
   isSignupModalOpen: state.login.isSignupModalOpen,
@@ -9,7 +9,7 @@ const mapStateToProps = state => ({
 const mapDispatchToProps = dispatch =>
   bindActionCreators(
     {
-      ...SignupModalActions,
+      ...signupModalActions,
     },
     dispatch,
   );
